+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.0.7
+
+- improve animation when textPainter.didExceedMaxLines = true
+- add HapticFeedback when text is animating
+
 ## 1.0.6
 
 - Fix StreamText text is short when textPainter.didExceedMaxLines = true